backend/shared/requirements.txt
aiohttp>=3.8.0
PyJWT>=2.0.0
fastapi>=0.100.0
pydantic>=2.0.0 